Add gmp compatibilty layer for imath
This commit adds a gmp compatibility layer for imath. The goal of
the compatibility layer is to provide the same interface and
functionality as gmp for a limited set of functions. The major
differences between the gmp api and the imath api that are addressed
by the compatiabilty layer are:

  1. gmp passes the output paramter as the first parmater
     imath passes the output parameter as the last paramter.

  2. gmp supports multiple rounding modes for division
     imath supports only the trucate rounding mode for division

  3. gmp support many operations that take an unsigned or a signed long
     imath supports only operations that take a signed long

In addition to these main differences, the compatibility layer also
fixes minor differences in behavior between imath and gmp functions
for certin inputs.

This commit also provides wrappers for import/export functions that
match the format provided by gmp. These functions can be used to
convert between gmp and imath integers.

A list of the 49 gmp rational and integer functions provided by this
compatibility layer are listed below.

Rational functions:
  mpq_clear
  mpq_cmp
  mpq_den_ref
  mpq_get_str
  mpq_init
  mpq_mul
  mpq_num_ref
  mpq_set
  mpq_set_ui
  mpq_sgn

Integer functions:
  mpz_abs
  mpz_add
  mpz_add_ui
  mpz_addmul
  mpz_cdiv_q
  mpz_clear
  mpz_cmp
  mpz_cmp_si
  mpz_cmpabs
  mpz_divexact
  mpz_divexact_ui
  mpz_divisible_p
  mpz_export
  mpz_fdiv_q
  mpz_fdiv_q_ui
  mpz_fdiv_r
  mpz_gcd
  mpz_get_si
  mpz_get_str
  mpz_get_ui
  mpz_import
  mpz_init
  mpz_lcm
  mpz_mul
  mpz_mul_2exp
  mpz_mul_ui
  mpz_neg
  mpz_pow_ui
  mpz_set
  mpz_set_si
  mpz_set_str
  mpz_set_ui
  mpz_sgn
  mpz_sizeinbase
  mpz_sub
  mpz_sub_ui
  mpz_submul
  mpz_swap
  mpz_tdiv_q