Feature: Add functions to return references to numer/denom
This commit adds two new functions to the rational number api:

  mp_rat_numer_ref()
  mp_rat_denom_ref()

These functions return a pointer to the underlying mp_int value used
by the rational number. It is the callers responsibility to not
modify the values in a way that invalidates any invariants needed by
the rational number implementation.

This function is needed for GMP compatiability.