BugFix: Change int to mp_small in api for mp_rat_set_value
This commit changes the function signature for mp_rat_set_value to
take an mp_small instead of an int. This change makes the function
consistent with the rest of the api which uses the mp_small typedef.
Note that mp_small is defined to be a long, so there should be no
change in functionality with this commit.