Force building imath with USE_LONG_LONG defined
This commit forces imath to be built with the -DUSE_LONG_LONG macro
defined. We explicitly define this value in the imath.h header
instead of relying on the value defined from the command line.

We force building this way so that the library is always built with
support for long long regardless of how it is compiled. This change
allows us to produce the same code for a library built using the
makefile and when imath is built when embedded in another project.