Improve mpz_get_str and mpq_get_str compatibility
This commit improves the gmp compatibility layer for the mpz_get_str
and mpq_get_str functions. There are two improvments:

  1. Use the same case as gmp when radix > 10
  2. Allow negative values for radix

Previously, imath would use uppercase for letters when creating a
string for a number in radix greater than 10. Now we use lowercase
for these letters to match gmp. We also support negative radix
values in the same way as gmp. A negative radix means to use
uppercase for any characters in the output string.

In addition to improved compatiability, we also updated the
gmp-compat-test suite to include tests for the mpz_get_str function.