Bug 8926: Use exact width int types for mp_digit and mp_word
This commit changes the data types used internally by imath to
implement large numbers. We now can use either the uint64_t or
uint32_t types for an mp_word.  The requirements for the mp_digit
and mp_word type is that mp_word is twice as large as an mp_digit.
The mp_digit type is used to store the individual digits of the
large integers.

By default we force the use of uint64_t for the mp_word type, but it
can be changed by modifying a #define in the imath.h header.

Note that we now have a dependency on the <stdint.h> header file. We
have also updated the makefile to build with -std=c99 flag because
the <stdint.h> header is part of the c99 standard.

The use of fixed width int types in <stdint.h> perhaps makes the
library less portable, but it increases the stability across
platforms because all versions will use the same size integers.

This change was motiviated by a build failure in suse linux that
occured because we were missing a #define for the max long long int
value ULLONG_MAX.