Fix undefined -MIN_LONG in mp_int_to_int.
Negating the minimal negative of a signed integer is undefined in C,
the result is not in the representable range. In two's complement,
the negation is the value itself.

Fix by negating as an unsigned integer, before casting to an signed
integers. "Negation" of an unsigned integer is defined a 2^n - x, where
n is the bit width and x the (unsigned, therefore non-negative) input
value.

0x8000000000000000 (or 0x80000000 if long is 32 bits) negated is the
value itself, then, after casting to a signed integer of the same size,
is interpreted as MIN_LONG.
The range check at the beginning of mp_int_to_int ensures that the
negative value is the only possible interpretation of that bit pattern.