Build with 64-bit words by default.
This reverts part of commit da9e923c, which forced the build to 64-bit words
regardless of the settings in the Makefile. Instead, define USE_32BIT_WORDS as
an exception, and default to 64-bit if that is not set. This should preserve
the existing behaviour for anyone using the old macro.