Cleanup: Use .tc for test files instead of .t.
Because there are so many test files, GitHub's linguist integration classifies
the repository as consisting mainly of "Terra" source. Obviously that's bogus,
but it annoys me, so let's change the extension to .tc.

I manually checked that .tc doesn't point to anything in the current public
config files for linguist. Of course, we won't really know whether this worked
until the stats update.