Remove the findsizes.py script.
This is no longer particularly useful, since we allowed ourselves access to
inttypes.h and thus have uintXX_t. Update the docs to clean up the mention of
the script and other size-related mumbo-jumbo.

I also dropped some cosmetic whitespace from elsewhere in doc.md, since it was
on the way.