imath: Replace accessor macros with inline functions.
We can't completely get rid of these names because they are part of the public
API, so this change preserves the macro-style MP_X names rather than using the
mp_x function style.

Unfortunately, we can't quite get rid of the macros entirely yet: They are used
internally as lvalues, which means a more complete fix will require rewriting
the expansion sites. So for now I've moved the macro definitions to imath.c.
This is also technically a breaking change, but clients were not meant to
depend on the lvalue usage, so pace http://www.hyrumslaw.com/ I'm dropping it.

Updates #29.