gmp_compat: Fix overflow in the uint conversion adapter.
The semantics of GMP mpz_get_ui require that if conversion fails out of range,
the return value is the low-order bits of the available value that will fit
into an unsigned long. However, the packing routine (get_long_bits) does not
take into account the magnitude of the given value.

For negative values of single-digit magnitude, conversion fails with MP_RANGE
because the value is negative: But in these cases, it is not the the magnitude
that's too large, it's the sign that's too small.

Nevertheless, get_long_bits assumes the magnitude is the problem and copies
digits based on how many fit into an unsigned long, not accounting for op->used
at all. When unsigned long is > 1 digit, it will try to grab 2 digits even if
the value only has 1 digit. And, because the digits at indices >= op->used are
garbage, it is getting uninitialized or left-over data.

In general this is a potential buffer overflow—but in the test cases
MP_ALLOC(op) is always big enough by happenstance.

Fortunately the fix is easy: Cap the number of digits to copy to the minimum of
the number available and the number that fit in an unsigned long.

Fixes #34.