imath: Drop switchable statics and stdio dependency.
Prior to this commit, defining DEBUG switched the static helper functions to
have external linkage. As far as I can tell nobody is making use of this.
Moreover, that switch exports two "print" functions that I used a lot when
debugging the initial library, but haven't done anything with in a long time.
Remove all that noise, and update the documentation.

Strictly speaking this is a breaking change, but only under DEBUG=Y, so I do
not feel it's necessary to protect it.