imath: Add documentation comments to the header.
Serves in partial satisfaction of #30. Each API function prototype in imath.h
has been given a Doxygen-style documentation comment, mostly based on the text
in doc.md but using indicative mood. I removed the telegraphic line-ending
comments that used to be here.

In making this change I found some errors in the comment text from doc.md, and
corrected them in transcription. However, I have not yet updated doc.md with
those changes; my intent is to remove that text entirely and generate it from
the comments added in this commit.