Add a make rule for doc.md from doc.md.in.
The doc.md file is now generated from doc.md.in, via a rule that uses the
mkdoc.py script.  The rule depend on the headers so that it will get updated
when they change.

As a sleazy hack, I made doc.md a dependency of the test target to ensure it
gets run. This will do till I set up better automation (see #27 and #28).

Serves in partial satisfaction of #30.