Add a Travis CI configuration for imath.
Updates #27. A very simple test plan that simply runs "make check" for each
interesting configuration. Right now the "interesting" configurations are macOS
(because I use it) and Linux (because everyone uses it), and the latter with
both Clang and GCC.

I had to update test.sh to be explicitly Bash rather than Bourne, since on
macOS "sh" is an alias for Bash but not so in Ubuntu.