Copy files into the build container.
Instead of mapping over the local filesystem, copy the build context into the
container and run the build there. The local state can still pollute the build
image, but make clean should take care of that, and this way we don't wind up
with .o files from the wrong architecture.